minor changes regarding comments in data_models
